Add RepaymentSchedule and RepaymentScheduleTest implementation
RepaymentSchedule represents the whole repayment schedule. It is initialized with the parameters loan amount, interest, starting repayment and a duration in years. After initialization a call to the calculate() method determines the remaining values for each month of the repayment schedule. These entries are stored within a list of RepaymentScheduleEntry.